Move validator and add some docstrings
